chore: refactor more stuff
