+++ src/libs/config.ts
@@ -1,0 +1,2 @@
+// TODO: actual scope TBD
+export const DEFAULT_VIEW_LAT_LONG = { lat: 22.4747061, lng: 88.3642162 };
